# order-microservice/src/main/kotlin/com/example/ordermicroservice/web/dto/ErrorDto.kt
package com.example.ordermicroservice.web.dto

import io.swagger.v3.oas.annotations.media.Schema

data class ErrorDto(
    @Schema(
        description = "Сообщение ошибки",
        example = "Сообщение об ошибке",
        required = true
    )
    val message: String
) {
}